requirements.txt
fastapi
sqlalchemy
uvicorn
pydantic
passlib
python-jose
bcrypt
python-multipart
google-generativeai
python-dotenv
easyocr 
pillow
pytest 
httpx 
pytest-asyncio
PyPDF2
pinecone
firebase_admin
python-dotenv
google-generativeai
langchain
langchain-google-genai
profanity-check